Reorganized master code
