+++ pkg/cmd/server/origin/handlers.go
@@ -1,0 +1,145 @@
+package origin
+
+import (
+	"encoding/json"
+	"net/http"
+
+	"bitbucket.org/ww/goautoneg"
+
+	restful "github.com/emicklei/go-restful"
+
+	kapi "github.com/GoogleCloudPlatform/kubernetes/pkg/api"
+	"github.com/GoogleCloudPlatform/kubernetes/pkg/apiserver"
+	"github.com/GoogleCloudPlatform/kubernetes/pkg/util"
+
+	"github.com/openshift/origin/pkg/api/latest"
+)
+
+// TODO We would like to use the IndexHandler from k8s but we do not yet have a
+// MuxHelper to track all registered paths
+func indexAPIPaths(handler http.Handler) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
+		if req.URL.Path == "/" {
+			// TODO once we have a MuxHelper we will not need to hardcode this list of paths
+			object := kapi.RootPaths{Paths: []string{
+				"/api",
+				"/api/v1beta3",
+				"/api/v1",
+				"/controllers",
+				"/healthz",
+				"/healthz/ping",
+				"/logs/",
+				"/metrics",
+				"/ready",
+				"/osapi",
+				"/osapi/v1beta3",
+				"/oapi",
+				"/oapi/v1",
+				"/swaggerapi/",
+			}}
+			// TODO it would be nice if apiserver.writeRawJSON was not private
+			output, err := json.MarshalIndent(object, "", "  ")
+			if err != nil {
+				http.Error(w, err.Error(), http.StatusInternalServerError)
+				return
+			}
+			w.Header().Set("Content-Type", restful.MIME_JSON)
+			w.WriteHeader(http.StatusOK)
+			w.Write(output)
+		} else {
+			handler.ServeHTTP(w, req)
+		}
+	})
+}
+
+func (c *MasterConfig) authorizationFilter(handler http.Handler) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
+		attributes, err := c.AuthorizationAttributeBuilder.GetAttributes(req)
+		if err != nil {
+			forbidden(err.Error(), "", w, req)
+			return
+		}
+		if attributes == nil {
+			forbidden("No attributes", "", w, req)
+			return
+		}
+
+		ctx, exists := c.RequestContextMapper.Get(req)
+		if !exists {
+			forbidden("context not found", attributes.GetAPIVersion(), w, req)
+			return
+		}
+
+		allowed, reason, err := c.Authorizer.Authorize(ctx, attributes)
+		if err != nil {
+			forbidden(err.Error(), attributes.GetAPIVersion(), w, req)
+			return
+		}
+		if !allowed {
+			forbidden(reason, attributes.GetAPIVersion(), w, req)
+			return
+		}
+
+		handler.ServeHTTP(w, req)
+	})
+}
+
+// cacheControlFilter sets the Cache-Control header to the specified value.
+func cacheControlFilter(handler http.Handler, value string) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
+		w.Header().Set("Cache-Control", value)
+		handler.ServeHTTP(w, req)
+	})
+}
+
+// namespacingFilter adds a filter that adds the namespace of the request to the context.  Not all requests will have namespaces,
+// but any that do will have the appropriate value added.
+func namespacingFilter(handler http.Handler, contextMapper kapi.RequestContextMapper) http.Handler {
+	infoResolver := &apiserver.APIRequestInfoResolver{util.NewStringSet("api", "osapi", "oapi"), latest.RESTMapper}
+
+	return http.HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
+		ctx, ok := contextMapper.Get(req)
+		if !ok {
+			http.Error(w, "Unable to find request context", http.StatusInternalServerError)
+			return
+		}
+
+		if _, exists := kapi.NamespaceFrom(ctx); !exists {
+			if requestInfo, err := infoResolver.GetAPIRequestInfo(req); err == nil {
+				// only set the namespace if the apiRequestInfo was resolved
+				// keep in mind that GetAPIRequestInfo will fail on non-api requests, so don't fail the entire http request on that
+				// kind of failure.
+
+				// TODO reconsider special casing this.  Having the special case hereallow us to fully share the kube
+				// APIRequestInfoResolver without any modification or customization.
+				namespace := requestInfo.Namespace
+				if (requestInfo.Resource == "projects") && (len(requestInfo.Name) > 0) {
+					namespace = requestInfo.Name
+				}
+
+				ctx = kapi.WithNamespace(ctx, namespace)
+				contextMapper.Update(req, ctx)
+			}
+		}
+
+		handler.ServeHTTP(w, req)
+	})
+}
+
+// If we know the location of the asset server, redirect to it when / is requested
+// and the Accept header supports text/html
+func assetServerRedirect(handler http.Handler, assetPublicURL string) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
+		if req.URL.Path == "/" {
+			accepts := goautoneg.ParseAccept(req.Header.Get("Accept"))
+			for _, accept := range accepts {
+				if accept.Type == "text" && accept.SubType == "html" {
+					http.Redirect(w, req, assetPublicURL, http.StatusFound)
+					return
+				}
+			}
+		}
+		// Dispatch to the next handler
+		handler.ServeHTTP(w, req)
+	})
+}
